[General] Rework time sync
The current hour, day and month are now saved to the world data and loaded back in when restarting the server.

The server's hour counter (renamed from timeCounter to hourCounter) is now set back to 0 when reaching a value of 24. However, reaching a value of 24 makes the day get incremented by 1 in the WorldInstance. If the day value exceeds the number of days in that particular month, the month value is incremented by 1 instead and the day is set to 1.

The /time command has been replaced by /sethour, /setday and /setmonth commands.