[General] Save certain data to unloaded cells if its origin is valid
More specifically, Container, ObjectState and ObjectLock packets that were sent as a result of clientside dialogue or global scripts will be saved to serverside cell data.

With the new container sync that requires all items to be stored in the serverside cell data before they can be removed by clients, quests adding items to containers such as the Thieves Guild's "Loot the Mages Guild" were broken. This commit fixes them.