[EventHandler] Don't try to exchange records for unloaded cells
There are only a few edge cases where this can be a problem. One example that comes to mind is when server scripts try to teleport newly logged in players to cells that don't actually exist, preventing the initial spawn cell (that they end up staying in) from actually being loaded by them.