Making guiID's great again
By pulling apart the creation of the enum table that is used by guiHelper it allows for other scripters to insert into the table guiHelper.names at the end and enum it again.
This allows for people to use it more dynamically and without conflicting guiID's.
I'm terrible at naming things. So you might want another name for the table.