[EventHandler] Stop filtering objects based on config.lua options
Previously, all the objects from an object packet were checked against the tables of disallowed object IDs in config.lua and filtered so they did not include any of them. Only a resulting table of acceptedObjects would then be passed to event validators and handlers.

It was an overcomplication that should be handled in a new type of hook called an event filter, if at all.

For now, the config.lua options with disallowed IDs for different packets are handled in default event validators instead, meaning that having one disallowed object in a packet makes the entire packet get rejected, which is fine for all of the affected events because their packets only ever have one object in them when sent by clients.