[Config] Improve instanced spawn and fix other minor issues
The script of the first door outside Sellus Gravius' office (the one next to the barrel with the ring) has also been disabled.

The instanced spawn now gives players their Release Identification papers.

soulValue has been added to the list of numerical record settings.

timestamps have been added to the playerKeyOrder that determines the formatting of JSON files for players.